setup-routes.sh: Set routing table only if it is not set
Now it doesn't matter if this skript is called to often, it always sets the rules correctly and every rule it sets is set exactly once.

The lines outside of the if have to be outside because when the VPN goes down the routes to the dev are cleared. But they do no harm, if they are executed to often because the kernel will notice the dublicate routes and reject them.