gave other ID to VPN-routingtable, small fix
Bird has fixed that routingtable 1 is the main routing table, so bird was unable to populate the VPN-routingtable. Now the VPN is table 10 and bird works happily.

We remote the br-fftr via VPN-Table-rule before setting it to avoid adding more and more dublicates of this rule when reexecuting the script.